fix: Dynamically import positions
