+++ apps/web/src/pages/balances/positions.tsx
@@ -3,14 +3,11 @@ import Head from 'next/head'
 
 import AssetsHeader from '@/components/balances/AssetsHeader'
 import { BRAND_NAME } from '@/config/constants'
-import DefiPositions from '@/features/positions'
-import { Box, Chip, Stack, Tooltip, Typography } from '@mui/material'
-import TotalAssetValue from '@/components/balances/TotalAssetValue'
-import usePositionsFiatTotal from '@/features/positions/hooks/usePositionsFiatTotal'
+import dynamic from 'next/dynamic'
+
+const DefiPositions = dynamic(() => import('@/features/positions'))
 
 const Positions: NextPage = () => {
-  const fiatTotal = usePositionsFiatTotal()
-
   return (
     <>
       <Head>
@@ -20,41 +17,6 @@ const Positions: NextPage = () => {
       <AssetsHeader />
 
       <main>
-        <Box mb={2}>
-          <TotalAssetValue fiatTotal={fiatTotal} title="Total positions value" />
-        </Box>
-
-        <Stack direction="row" alignItems="center" gap={1}>
-          <Typography variant="h4" fontWeight={700}>
-            Positions
-          </Typography>
-          <Tooltip
-            title="Experimental. Data may be missing or outdated."
-            placement="top"
-            arrow
-            slotProps={{
-              tooltip: {
-                sx: {
-                  maxWidth: { xs: '250px', sm: 'none' },
-                },
-              },
-            }}
-          >
-            <Chip label="Beta" size="small" sx={{ backgroundColor: 'background.lightGrey', letterSpacing: '0.4px' }} />
-          </Tooltip>
-        </Stack>
-
-        <Box mb={1}>
-          <Typography
-            variant="caption"
-            sx={{
-              color: 'text.secondary',
-              letterSpacing: '1px',
-            }}
-          >
-            Position balances are not included in the total asset value.
-          </Typography>
-        </Box>
         <DefiPositions />
       </main>
     </>
